+++ CHANGELOG.md
@@ -15,7 +15,8 @@ For inspiration and motivation, see [Keep a CHANGELOG](https://keepachangelog.co
     - Many smaller documentation improvements from multiple people looking closely at docs.
     - Started official documenation for integration tests.
 - Fly.io:
-    - Assumes one fly app exists; default to region 'sea'.
+    - Assumes you have no existing Fly apps.
+    - Identifies lowest-latency region to deploy to; defaults to 'sea' if that information is unattainable.
     - Deprecate use of `flyctl`; use `fly` consistently throughout.
 - Heroku:
     - Uses `'*'` for `ALLOWED_HOSTS` on Heroku, as a temp fix.
@@ -36,6 +37,7 @@ For inspiration and motivation, see [Keep a CHANGELOG](https://keepachangelog.co
     - Validates pytest call, to run either unit tests or integration tests, not both. Also require `-s` for integration tests.
     - New tool for standing up a dev environment: `build_dev_env.py`
     - Started less formal notes about each platform, in *developer_resources/*.
+    - On Fly deployments, updates `fly open` calls to `fly apps open -a <app-name>`. Also updates deprecated `fly regions list -a` to get region with lowest latency.
 
 
 ### 0.5.15